+++ HSCore-config/HSCore-config-common/src/main/java/me/hsgamer/hscore/config/path/LongConfigPath.java
