+++ HSCore-config/HSCore-config-yaml/pom.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xmlns="http://maven.apache.org/POM/4.0.0"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <artifactId>HSCore-config</artifactId>
+    <groupId>me.HSGamer</groupId>
+    <version>1.6-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+
+  <artifactId>HSCore-config-yaml</artifactId>
+
+  <dependencies>
+    <dependency>
+      <groupId>me.HSGamer</groupId>
+      <artifactId>HSCore-config-common</artifactId>
+      <version>1.6-SNAPSHOT</version>
+    </dependency>
+    <dependency>
+      <groupId>me.carleslc.Simple-YAML</groupId>
+      <artifactId>Simple-Yaml</artifactId>
+      <version>1.7.1</version>
+    </dependency>
+  </dependencies>
+
+</project>
